+++ test/acceptance/export/csv.js
@@ -37,7 +37,7 @@ test('CSV format', function(done){
         var row1 = rows[1].split(',');
 
         assert.equal(row0[1], 'created_at');
-        assert.equal(row1[1], '2011-09-21 14:02:21');
+        assert.equal(row1[1], '2011-09-21 14:02:21.314252');
 
         done();
     });
@@ -130,7 +130,7 @@ test('GET /api/v1/sql as csv', function(done){
         method: 'GET'
     },{ }, function(res){
         assert.equal(res.statusCode, 200, res.body);
-        var body = "cartodb_id,geom\r\n1,SRID=4326;POINT(-3.699732 40.423012)";
+        var body = 'cartodb_id,geom\r\n1,"SRID=4326;POINT(-3.699732 40.423012)"\r\n';
         assert.equal(body, res.body);
         done();
     });
@@ -157,10 +157,30 @@ test('GET /api/v1/sql as csv, properly escaped', function(done){
         method: 'GET'
     },{ }, function(res){
         assert.equal(res.statusCode, 200, res.body);
-        var body = 'cartodb_id,address\r\n1,"Calle de Pérez Galdós 9, Madrid, Spain"';
+        var body = 'cartodb_id,address\r\n1,"Calle de Pérez Galdós 9, Madrid, Spain"\r\n';
         assert.equal(body, res.body);
         done();
     });
+});
+
+test('GET /api/v1/sql as csv, concurrently', function(done){
+
+    var concurrency = 4;
+    var waiting = concurrency;
+    for (var i=0; i<concurrency; ++i) {
+
+      assert.response(app, {
+          url: '/api/v1/sql?q=SELECT%20cartodb_id,%20address%20FROM%20untitle_table_4%20LIMIT%201&format=csv',
+          headers: {host: 'vizzuality.cartodb.com'},
+          method: 'GET'
+      },{ }, function(res){
+          assert.equal(res.statusCode, 200, res.body);
+          var body = 'cartodb_id,address\r\n1,"Calle de Pérez Galdós 9, Madrid, Spain"\r\n';
+          assert.equal(body, res.body);
+          if ( ! --waiting ) done();
+      });
+
+    }
 });
 
 });